Added IsolationForest as well as the detection architecture

+++ src/main/java/fr/nelpats/model/detection/algorithm/Algorithm.java
@@ -5,8 +5,8 @@ public abstract class Algorithm {
     public abstract int getDelay();
 
 
-    public int[] getDelaySample(int n) {
-        int[] result = new int[n];
+    public double[] getDelaySample(int n) {
+        double[] result = new double[n];
 
         for (int i = 0; i < n; i++) {
             result[i] = getDelay();
